requirements.txt
Django==4.1.6
django-tinymce==3.5.0
django-crispy-forms==1.14.0
fontawesomefree==6.3.0
django-recaptcha==3.0.0
six